function that prints all the elements of a listint_t list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,6 +1,6 @@
 #include "lists.h"
 #include <stdio.h>
-#include <stddef.h>
+
 /**
  * print_listint - prints all the elements of a listint_t list
  * @h: a pointer to the start of the list
@@ -10,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-unsigned int i = 0;
+size_t i = 0;
 while (h != NULL)
 {
 printf("%d\n", h->n);
